fix 3.0.0 bindgen
