+++ openssl-sys/src/handwritten/evp.rs
@@ -505,30 +505,6 @@ extern "C" {
     ) -> c_int;
 }
 
-cfg_if! {
-    if #[cfg(ossl300)] {
-        extern "C" {
-            pub fn EVP_PKEY_CTX_set_hkdf_mode(ctx: *mut EVP_PKEY_CTX, mode: c_int) -> c_int;
-            pub fn EVP_PKEY_CTX_set_hkdf_md(ctx: *mut EVP_PKEY_CTX, md: *const EVP_MD) -> c_int;
-            pub fn EVP_PKEY_CTX_set1_hkdf_salt(
-                ctx: *mut EVP_PKEY_CTX,
-                salt: *const u8,
-                saltlen: c_int,
-            ) -> c_int;
-            pub fn EVP_PKEY_CTX_set1_hkdf_key(
-                ctx: *mut EVP_PKEY_CTX,
-                key: *const u8,
-                keylen: c_int,
-            ) -> c_int;
-            pub fn EVP_PKEY_CTX_add1_hkdf_info(
-                ctx: *mut EVP_PKEY_CTX,
-                info: *const u8,
-                infolen: c_int,
-            ) -> c_int;
-        }
-    }
-}
-
 const_ptr_api! {
     extern "C" {
         pub fn EVP_PKCS82PKEY(p8: #[const_ptr_if(any(ossl110, libressl280))] PKCS8_PRIV_KEY_INFO) -> *mut EVP_PKEY;
