+++ openssl-sys/src/handwritten/mod.rs
@@ -8,6 +8,8 @@ pub use handwritten::dsa::*;
 pub use handwritten::ec::*;
 pub use handwritten::evp::*;
 pub use handwritten::hmac::*;
+pub use handwritten::kdf::*;
+pub use handwritten::ocsp::*;
 pub use handwritten::pkcs7::*;
 pub use handwritten::rsa::*;
 pub use handwritten::sha::*;
@@ -27,6 +29,8 @@ mod dsa;
 mod ec;
 mod evp;
 mod hmac;
+mod kdf;
+mod ocsp;
 mod pkcs7;
 mod rsa;
 mod sha;
